refactor(model_connectors)!: move load connectors into own directory
BREAKING CHANGE: model connectors are at a different path in package
