+++ geojson_modelica_translator/model_connectors/load_connectors/base.py
@@ -52,7 +52,7 @@ class Base(object):
         self.system_parameters = system_parameters
 
         # initialize the templating framework (Jinja2)
-        self.template_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "templates")
+        self.template_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "templates")
         self.template_env = Environment(loader=FileSystemLoader(searchpath=self.template_dir))
 
         # store a list of the templated files to include when building the package
