+++ geojson_modelica_translator/model_connectors/load_connectors/teaser_ets_coupling.py
@@ -32,7 +32,7 @@ import glob
 import os
 import shutil
 
-from geojson_modelica_translator.model_connectors.base import \
+from geojson_modelica_translator.model_connectors.load_connectors.base import \
     Base as model_connector_base
 from geojson_modelica_translator.modelica.input_parser import PackageParser
 from geojson_modelica_translator.utils import ModelicaPath, copytree
