+++ tests/geojson_modelica_translator/test_translator.py
@@ -33,11 +33,11 @@ import os
 from geojson_modelica_translator.geojson_modelica_translator import (
     GeoJsonModelicaTranslator
 )
-from geojson_modelica_translator.model_connectors.spawn import \
+from geojson_modelica_translator.model_connectors.load_connectors.spawn import \
     SpawnConnector as spawn_load
-from geojson_modelica_translator.model_connectors.teaser import \
+from geojson_modelica_translator.model_connectors.load_connectors.teaser import \
     TeaserConnector as teaser_load
-from geojson_modelica_translator.model_connectors.time_series import \
+from geojson_modelica_translator.model_connectors.load_connectors.time_series import \
     TimeSeriesConnector as timeseries_load
 from geojson_modelica_translator.system_parameters.system_parameters import (
     SystemParameters
