+++ tests/model_connectors/test_district_heating_system.py
@@ -36,7 +36,7 @@ from pathlib import Path
 from geojson_modelica_translator.geojson_modelica_translator import (
     GeoJsonModelicaTranslator
 )
-from geojson_modelica_translator.model_connectors.base import \
+from geojson_modelica_translator.model_connectors.load_connectors.base import \
     Base as model_connector_base
 from geojson_modelica_translator.model_connectors.district_system import (
     DistrictSystemConnector
