+++ tests/model_connectors/test_district_system.py
@@ -33,7 +33,7 @@ import os
 from geojson_modelica_translator.geojson_modelica_translator import (
     GeoJsonModelicaTranslator
 )
-from geojson_modelica_translator.model_connectors.base import \
+from geojson_modelica_translator.model_connectors.load_connectors.base import \
     Base as model_connector_base
 from geojson_modelica_translator.model_connectors.district_system import (
     DistrictSystemConnector
