+++ tests/model_connectors/test_teaser.py
@@ -34,7 +34,7 @@ import os
 from geojson_modelica_translator.geojson_modelica_translator import (
     GeoJsonModelicaTranslator
 )
-from geojson_modelica_translator.model_connectors.teaser import TeaserConnector
+from geojson_modelica_translator.model_connectors.load_connectors.teaser import TeaserConnector
 from geojson_modelica_translator.system_parameters.system_parameters import (
     SystemParameters
 )
