+++ tests/model_connectors/test_time_series.py
@@ -33,7 +33,7 @@ import os
 from geojson_modelica_translator.geojson_modelica_translator import (
     GeoJsonModelicaTranslator
 )
-from geojson_modelica_translator.model_connectors.time_series import (
+from geojson_modelica_translator.model_connectors.load_connectors.time_series import (
     TimeSeriesConnector
 )
 from geojson_modelica_translator.system_parameters.system_parameters import (
