+++ tests/model_connectors/test_time_series_mft_ets.py
@@ -33,7 +33,7 @@ import os
 from geojson_modelica_translator.geojson_modelica_translator import (
     GeoJsonModelicaTranslator
 )
-from geojson_modelica_translator.model_connectors.time_series_mft_ets_coupling import (
+from geojson_modelica_translator.model_connectors.load_connectors.time_series_mft_ets_coupling import (
     TimeSeriesConnectorMFTETS
 )
 from geojson_modelica_translator.system_parameters.system_parameters import (
